[REF] l10n_ve_accountant, l10n_ve_sale: reorganizar lógica de impuestos y robustecer cálculo
Se realizan los siguientes cambios para mejorar la modularidad y estabilidad:
- Se elimina la herencia de líneas de compra y venta en l10n_ve_accountant.
- Se reubica la lógica de preparación de impuestos de ventas en l10n_ve_sale.
- En account_tax, se añade una validación mediante hasattr para asegurar que
  el cálculo de totales de impuestos (tax_totals) solo intente procesar líneas
  si los módulos que gestionan los campos de moneda extranjera están instalados,
  previniendo errores de ejecución en entornos con instalaciones parciales.
- Se ajusta la obtención de la compañía usando el registro activo para mayor precisión.

References: https://binaural.odoo.com/odoo/action-341/67612
