+++ l10n_ve_accountant/models/__init__.py
@@ -1,8 +1,6 @@
 from . import res_company
 from . import account_move
 from . import account_move_line
-from . import purchase_order_line
-from . import sale_order_line
 from . import account_tax
 from . import account_partial_reconcile
 from . import account_payment
